+++ app/i18n/translations/whitelist_dev.json
@@ -4,12 +4,7 @@
   "ln.invoicesTab.invoice.creationDate",
   "ln.invoicesTab.invoice.settleDate",
   "ln.paymentsTab.payment.creationDate",
-  "ticket.timestamp",
   "heatmap.month",
-  "heatmap.tooltip.vote",
-  "heatmap.tooltip.live",
-  "heatmap.tooltip.revoked",
-  "heatmap.tooltip.maturing",
   "heatmap.tooltip.date",
   "txDetails.timestamp",
   "snackbar.defaults.error",
